[Pipe] Refactor convert_to_balance under non-test package.
Since fairscale.nn.Pipe still uses 'balance' and 'devices' parameters,
other frameworks like fairseq still use these parameters. As a result, the
`convert_to_balance` method is a nice utility to use for migrating to PyTorch
Pipe without changing a lot of code in other frameworks.

In addition to this I've renamed the method to be more illustrative of what it
does and also allowed an optional devices parameter.

Differential Revision: [D25987273](https://our.internmc.facebook.com/intern/diff/D25987273/)

ghstack-source-id: 120086065
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50860