Fix CUDA RPC Stream Synchronization
When converting RPC Message into Python objects, we were not using
a CUDAFuture for the chained Future. As a result, the streams are
not synchronized when calling `rpc_async(...).wait()`. This commit
uses `Future::then` API to create the chained Future, which will
be creating a CUDAFuture if the existing Future is a CUDA one.

fixes #50881
fixes #50839

ghstack-source-id: db34a5763a63f1d660c61552e5d3c71ae52e5875
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50949